src/main/java/com/trendyol/shoppingcard/controller/CategoryController.java
package com.trendyol.shoppingcard.controller;

import com.trendyol.shoppingcard.common.ApiResponse;
import com.trendyol.shoppingcard.model.Category;
import com.trendyol.shoppingcard.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/category")
public class CategoryController {

    private CategoryService categoryService;

    @Autowired
    public CategoryController(CategoryService categoryService) {
        this.categoryService = categoryService;
    }

    @PostMapping("/save")
    public ApiResponse<Category> save(@RequestBody Category category) {
        return ApiResponse.response(categoryService.save(category));
    }

    @GetMapping("/")
    public ApiResponse<Category> get(String categoryId) {
        return ApiResponse.response(categoryService.get(categoryId));
    }
}
